Rename 'Common' to 'BuildingBlock' in SQL constants.
