+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/command/query/CreateCommandTableQuery.java
@@ -22,10 +22,10 @@ package org.spine3.server.storage.jdbc.command.query;
 
 import org.spine3.server.storage.jdbc.query.CreateTableQuery;
 
-import static org.spine3.server.storage.jdbc.Sql.Common.BRACKET_CLOSE;
-import static org.spine3.server.storage.jdbc.Sql.Common.BRACKET_OPEN;
-import static org.spine3.server.storage.jdbc.Sql.Common.COMMA;
-import static org.spine3.server.storage.jdbc.Sql.Common.SEMICOLON;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.BRACKET_CLOSE;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.BRACKET_OPEN;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.COMMA;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.SEMICOLON;
 import static org.spine3.server.storage.jdbc.Sql.Query.CREATE_IF_MISSING;
 import static org.spine3.server.storage.jdbc.Sql.Query.PRIMARY_KEY;
 import static org.spine3.server.storage.jdbc.Sql.Type.BLOB;
