+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/event/query/FilterAndSortQuery.java
@@ -36,13 +36,13 @@ import java.sql.PreparedStatement;
 import java.util.Iterator;
 
 import static org.spine3.base.Identifiers.idToString;
-import static org.spine3.server.storage.jdbc.Sql.Common.BRACKET_CLOSE;
-import static org.spine3.server.storage.jdbc.Sql.Common.BRACKET_OPEN;
-import static org.spine3.server.storage.jdbc.Sql.Common.COMMA;
-import static org.spine3.server.storage.jdbc.Sql.Common.EQUAL;
-import static org.spine3.server.storage.jdbc.Sql.Common.GT;
-import static org.spine3.server.storage.jdbc.Sql.Common.LT;
-import static org.spine3.server.storage.jdbc.Sql.Common.SEMICOLON;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.BRACKET_CLOSE;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.BRACKET_OPEN;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.COMMA;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.EQUAL;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.GT;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.LT;
+import static org.spine3.server.storage.jdbc.Sql.BuildingBlock.SEMICOLON;
 import static org.spine3.server.storage.jdbc.Sql.Query.AND;
 import static org.spine3.server.storage.jdbc.Sql.Query.ASC;
 import static org.spine3.server.storage.jdbc.Sql.Query.OR;
